Move things around preparing for RPC out of process
