+++ CMakeLists.txt
@@ -292,6 +292,7 @@ add_subdirectory(nano/crypto_lib)
 add_subdirectory(nano/secure)
 add_subdirectory(nano/lib)
 add_subdirectory(nano/node)
+add_subdirectory(nano/rpc)
 add_subdirectory(nano/nano_node)
 
 if (NANO_TEST OR RAIBLOCKS_TEST)
