+++ nano/lib/rpcconfig.hpp
@@ -10,17 +10,16 @@ namespace nano
 class jsonconfig;
 
 /** Configuration options for RPC TLS */
-class rpc_secure_config
+class rpc_secure_config final
 {
 public:
-	rpc_secure_config ();
 	nano::error serialize_json (nano::jsonconfig &) const;
 	nano::error deserialize_json (nano::jsonconfig &);
 
 	/** If true, enable TLS */
-	bool enable;
+	bool enable{ false };
 	/** If true, log certificate verification details */
-	bool verbose_logging;
+	bool verbose_logging{ false };
 	/** Must be set if the private key PEM is password protected */
 	std::string server_key_passphrase;
 	/** Path to certificate- or chain file. Must be PEM formatted. */
@@ -33,10 +32,10 @@ public:
 	std::string client_certs_path;
 };
 
-class rpc_config
+class rpc_config final
 {
 public:
-	rpc_config (bool = false);
+	explicit rpc_config (bool = false);
 	nano::error serialize_json (nano::jsonconfig &) const;
 	nano::error deserialize_json (bool & upgraded_a, nano::jsonconfig &);
 	nano::network_constants network_constants;
